restructure cypress test code
