+++ cypress/e2e/Authentication/LogStaffOut/main.js
