--- cypress/integration/common/globalBefore.js
@@ -1,3 +1,0 @@
-before(() => {
-  // Code in this hook really runs once before all tests
-})
